[af-207] refactor: Product alcohol data type 수정

+++ src/main/java/com/drunkenlion/alcoholfriday/domain/admin/product/dto/ProductCreateRequest.java
@@ -30,7 +30,7 @@ public class ProductCreateRequest {
     private Long quantity;
 
     @Schema(description = "술 도수")
-    private Long alcohol;
+    private Double alcohol;
 
     @Schema(description = "제품 재료")
     private String ingredient;
